+++ inquire/src/terminal/crossterm.rs
@@ -135,7 +135,7 @@ impl Terminal for CrosstermTerminal {
         self.get_writer().flush()
     }
 
-    fn get_size(&self) -> Result<super::TerminalSize> {
+    fn get_size(&self) -> Result<Option<super::TerminalSize>> {
         terminal::size().map(|(width, height)| super::TerminalSize::new(width, height))
     }